requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic>=2.0.0
PyYAML==6.0.1
prometheus-client==0.19.0
supervision==0.18.0
bytetrack==0.1.0
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1

